API

Modules

Types

SurvivalAnalysis.ConcordanceWeights — Type

ConcordanceWeights(S::Int8, G::Int8, tied_preds::Float64 tied_times::Float64)

Weights used in the concordance function. S and G reflect the power applied to the Kaplan-Meier estimates of the survival and censoring distributions of the fitted data used to apply weighting at a given time. For example ConcordanceWeights(1, -2, 0.5, 0.5) will tell the concordance function to multiply concordant pairs at time t by S(t)/G(t)². tied_preds and tied_times determine how to handle ties in predictions and observed times respectively.

See concordance for full examples.

Note❗ It is strongly recommended that S ≥ 0 and G ≤ 0.

SurvivalAnalysis.Surv — Method

Surv(start, stop)
Surv(time)
Surv(time, status, type)

This is the entry point object into survival modelling.

SurvivalAnalysis.SurvivalModel — Type

SurvivalModel

Abstract type for all models implemented in, or extending, this package. Type 'inherits'

from JuliaStats.StatisticalModel to enable formula fitting and predicting interface.

SurvivalAnalysis.SurvivalPrediction — Method

SurvivalPrediction(;
    distr::Union{Nothing, Vector{<:Distribution}} = nothing,
    lp::Union{Nothing, Vector{T}}  = nothing,
    crank::Union{Nothing, Vector{T}} = nothing,
    time::Union{Nothing, Vector{T}} = nothing,
    fit_times::Union{Nothing, Vector{T}} = nothing,
    survival_matrix::Union{Matrix{T}, Nothing} = nothing
    ) where {T<:Number}

Survival models can make multiple types of predictions including:

distr - A survival time distribution (implemented with Distributions.jl)

lp - A linear predictor (usually Xβ, i.e. covariates * fitted coefficients)

crank - A generic continuous relative risk ranking

time - A survival time

survival_matrix - A matrix of predicted survival probabilities where rows are

observations and columns are fitted survival times corresponding to fit_times.

These predictions can only exist in a finite number of combinations so they are aggregated in types within this package (and automatically determined within this function):

DeterministicSurvivalPrediction(lp, crank, time)

DiscreteSurvivalPrediction(distr, lp, crank, time, survival_matrix)

ContinuousSurvivalPrediction(distr, lp, crank, time)

Absolutely no transformations take require assumptions take place within this function but it does transform survival_matrix to a Distributions.DiscreteNonParametric and vice versa.

Examples

julia> SurvivalPrediction(time = randn(5)); # DeterministicSurvivalPrediction

julia> SurvivalPrediction(fit_times = randn(5), survival_matrix = randn((2, 5))); # DiscreteSurvivalPrediction

julia> SurvivalPrediction(distr = fill(Exponential(), 2)); # ContinuousSurvivalPrediction

Functions

RecipesBase.apply_recipe — Function

plot(npe::SurvivalEstimator, plot_confint::Bool = true; level = 0.95)
plot(npe::StatsModels.TableStatisticalModel{KaplanMeier, Matrix{Float64}},
    plot_confint::Bool = true; level = 0.95)

Recipe for plotting fitted non-parametric estimators, `npe`. If `plot_confint` then
confidence intervals also plotted at a `level`% confidence level.

# Examples
```jldoctest
julia> using Plots

julia> data = DataFrame(t = randn(10), d = [trues(5)..., falses(5)...]);

julia> plot(kaplan_meier(@formula(Srv(t, d) ~ 1), data));

julia> plot(nelson_aalen(@formula(Srv(t, d) ~ 1), data).model);
```

SurvivalAnalysis.Srv — Function

Srv(T::Symbol, Δ::Symbol, type::Int = 1)

Create a `SurvTerm` object for internal use for fitting models. Arguments, `T` and `Δ`
should be a reference to the name of the time and status variables in a `DataFrame`
respectively. `type` should be `1` (or omitted) for right-censoring or `-1`
for left-censoring.

# Examples
```jldoctest
julia> using DataFrames

julia> data = DataFrame(t = [1, 2, 3], d = [1, 0, 0])

julia> @formula(Srv(t, d) ~ 1)
FormulaTerm
Response:
(t,d)->Srv(t, d)
Predictors:
1

julia> Srv(t, d, -1)
Response:
(t,d)->Srv(t, d, -1)
Predictors:
1
```

SurvivalAnalysis.concordance — Function

concordance(
    truth::OneSidedSurv, prediction::Vector{<:Number}, weights::Union{Symbol, ConcordanceWeights};
    tied_preds=0.5, tied_times=0, cutoff=nothing, train::OneSidedSurv=nothing, rev=false
)
cindex(
    truth::OneSidedSurv, prediction::Vector{<:Number}, weights::Union{Symbol, ConcordanceWeights};
    tied_preds=0.5, tied_times=0, cutoff=nothing, train::OneSidedSurv=nothing, rev=false
)

Generic function to call any concordance index method. Concordance is a measure of discrimination which evaluates if a prediction is concordant with the truth, i.e. let Tᵢ,Tⱼ be the true survival times, truth, for observations i and j and let ϕᵢ, ϕⱼ be predictions, prediction, then these are concordant if $ϕᵢ > ϕⱼ ⟺ Tᵢ > Tⱼ$.

In survival analysis the generic C-index is defined as follows

\[C = \frac{\sum_{i≠j} W(tᵢ)I(tᵢ < tⱼ, yᵢ < yⱼ, tᵢ < τ)δᵢ}{W(tᵢ)I(tᵢ < tⱼ, tᵢ < τ)δᵢ}\]

where tᵢ,tⱼ are true survival times, yᵢ,yⱼ are predictions, τ is a cutoff time used to ensure stability even when censoring is high, δᵢ is the censoring indicator and W is a weighting function determined by weights as follows:

* `:I` or `:Harrell` - Harrell's C - ``W(tᵢ) = 1``
* `:G2` or `:Uno` - Uno's C - ``W(tᵢ) = 1/G(tᵢ)``
* `:SG` or `:Schemper` - Schemper's C - ``W(tᵢ) = S(tᵢ)/G(tᵢ)``
* `:S` or `:Peto` - Peto-Wilcoxon's C - ``W(tᵢ) = S(tᵢ)``

where S(tᵢ) and G(tᵢ) are respectively the Kaplan-Meier estimates of the survival and censoring distributions of the training data at time tᵢ. For any other combination of weights pass a ConcordanceWeights object to weights. Note❗ If training data is not provided to train then truth is used to estimate S and G but it is strongly recommended to provide the training data if possibe.

We also include an implementation of Gönen-Heller's C with weights = :GH or :Gonen. Note❗ this is actually a very different method from the others and calculates concordance for predictions from a Cox PH model only. We may move this to its own function in the future to avoid misuse.

There is open debate about how to handle ties when calculating the concordance. Ties can occur in predictions and in the observed survival times. The defaults here tied_preds=0.5 and tied_times=0 are set as these seem to be the most common but they can be changed. Note❗ If you pass a ConcordanceWeights object then the tied weights specified in this will take priority.

Note❗ For predictions from PH models or any model where the prediction represents a relative risk then a higher value of ϕ implies a higher risk of event which will result in a lower survival time. In this case a prediction is concordant with the survival time if $ϕᵢ < ϕⱼ ⟺ Tᵢ > Tⱼ$. To do this within the function just set rev=true.

julia> T = [1.0,0.1,pi,0.9,0.4,20,1,5,9,2.5];

julia> ϕ = [0.1,0.2,0.1,0.9,0.25,exp(2),8,log(9),2,8];

julia> Δ = [true,true,false,true,true,false,true,false,true,true];

julia> Y = Surv(T, Δ, :r);

julia> train = Surv([3,3,9,12,2,1,8,4,10,8], Δ, :r);

julia> concordance(Y, ϕ, cutoff = threshold_risk(Y, 0.8)) # Harrell's C cutoff when 80% data is censored or dead
SurvivalAnalysis.Concordance

Harrell's C = 0.6153846153846154

Cutoff: T ≤ 9.0
Counts:
 Pairs  Comparable  Concordant  Disconcordant
    90          39          23             14
Weights:
 IPCW  Tied preds  Tied times
    1         0.5         0.0
Ties:
 Times  Preds  Both
     1      2     0
Weighted calculation:
 Numerator  Denominator         C
      24.0         39.0  0.615385

julia> concordance(Y, ϕ, :Uno, train=train) # Uno's C
SurvivalAnalysis.Concordance

Uno's C = 0.6221374045801525

Cutoff: T ≤ 20.0
Counts:
 Pairs  Comparable  Concordant  Disconcordant
    90          39          23             14
Weights:
 IPCW  Tied preds  Tied times
 1/G²         0.5         0.0
Ties:
 Times  Preds  Both
     1      2     0
Weighted calculation:
 Numerator  Denominator         C
   28.1728       45.284  0.622137

julia> cindex(Y, ϕ, ConcordanceWeights(5, -3, 0.5, 0.5, "Silly Weights"); train=train) # Custom weights
SurvivalAnalysis.Concordance

Silly Weights C = 0.6070334788405888

Cutoff: T ≤ 20.0
Counts:
 Pairs  Comparable  Concordant  Disconcordant
    90          40          23             14
Weights:
  IPCW  Tied preds  Tied times
 S⁵/G³         0.5         0.5
Ties:
 Times  Preds  Both
     1      2     0
Weighted calculation:
 Numerator  Denominator         C
   25.6265       42.216  0.607033

SurvivalAnalysis.cum_hazard — Method

cum_hazard(d::UnivariateDistribution, x::Real)
Hₜ(d::UnivariateDistribution, x::Real)

Compute the cumulative hazard function of distribution d at point x.

The cumulative hazard function for random variable $t$ is defined as

\[H_t(x) = \int^x_0 h_t(u) du = -log(S_t(x))\]

where $hₜ$ is the hazard function of $t$ and $Sₜ$ is the survival function of $t$.

Examples

julia> using Distributions

julia> cum_hazard(Binomial(5, 0.5), 3)
1.6739764335716711

SurvivalAnalysis.hazard — Method

hazard(d::UnivariateDistribution, x::Real)
hₜ(d::UnivariateDistribution, x::Real)

Compute the hazard function of distribution d at point x.

The hazard function for random variable $t$ is defined as

\[h_t(x) = \frac{f_t(x)}{S_t(x)}\]

where $fₜ$ is the pdf of $t$ and $Sₜ$ is the survival function of $t$.

Examples

julia> using Distributions

julia> hazard(Binomial(5, 0.5), 3)
1.6666666666666679

Macros and Constants