API

Surv

SurvivalAnalysis.Surv — Type

Surv
OneSidedSurv <: Surv
TwoSidedSurv <:Surv
RCSurv <: OneSidedSurv
LCSurv <: OneSidedSurv
IntSurv <: TwoSidedSurv

Surv is the abstract type for all survival outcome representations in this package. See Surv for constructing a Surv object.

Available methods:

length - Get length of a survival object

outcome_times - Get times at which an outcome (event or censoring) takes place

event_times - Get times at which an event (not censoring) takes place

outcome_status - Get vector of survival indicators

unique_outcome_times - Get unique times at which an outcome takes place

unique_event_times - Get unique times at which an event takes place

total_events - Get total number of events (optionally at a given time)

total_censored - Get total number censored (optionally at a given time)

total_outcomes - Get total number of outcomes (optionally at a given time)

total_risk - Get total number at risk (optionally at a given time)

surv_stats - Get set of useful summary statistics

threshold_risk - Get the time at which a given proportion of observations are no longer at risk

merge - Merge survival objects

reverse - Reverse survival outcome (not in-place)

SurvivalAnalysis.Surv — Method

Surv(start, stop)
Surv(time)
Surv(time, status, type)

Survival analysis is dependent on representing both known and unknown survival time outcomes for an observation. When the true survival time is unknown we say an observation is censored and we instead record their censoring time. Let $(T, Δ)$ be a one-sided survival outcome, then $T$ is the outcome time (the time at which an outcome, event or censoring, is observed) and $Δ$ is the survival indicator (1 if the outcome is an event and 0 if censoring). For example, if a patient drops out of a study at time 5, then they are recorded as $(T=5, Δ=0)$. Two-sided survival outcomes are more simply recorded as $(T₁,T₂)$ which means the true event time was somewhere between $T₁$ and $T₂$.

There are three censoring types:

Right (Surv(time, status, :r) - The true outcome time occurs at some time after the observed censoring time

Left (Surv(time, status, :l)) - The true outcome time occurs at some time before the observed censoring time

Interval (Surv(start, stop)) - The true outcome time occurs at some time within the observed censoring times

If no status vector is passed to the function then it is assumed no-one is censored - this is rarely useful in practice as in this case one could simply use regression models.

Note❗ Whilst this package supports functionality for all censoring types, currently only methods for right censoring are included.

Examples

julia> Surv([1, 2, 3], [true, false, true], :r) # right-censoring
["1.0", "2.0+", "3.0"]

julia> Surv([1, 2, 3], [1, 0, 1], :l) # left-censoring
["1.0", "2.0-", "3.0"]

julia> Surv([1, 2, 3], [5, 6, 7]) # interval-censoring
[(1.0, 5.0), (2.0, 6.0), (3.0, 7.0)]

Base.length — Method

length(srv::OneSidedSurv)
length(srv::TwoSidedSurv)

Get length of a survival object.

Examples

julia> length(Surv([1, 2, 3], [true, false, true], :r))
3

julia> length(Surv([1, 2, 3], [5, 6, 7]))
3

Base.merge — Method

merge(A::OneSidedSurv...)
merge(A::TwoSidedSurv...)

Merge survival objects.

Examples

julia> srv1 = Surv([1, 2, 3], [false, true, false], :r)
["1.0+", "2.0", "3.0+"]

julia> srv2 = Surv([4, 5, 6], [false, true, true], :r)
["4.0+", "5.0", "6.0"]

julia> merge(srv1, srv2) # OneSidedSurv
["1.0+", "2.0", "3.0+", "4.0+", "5.0", "6.0"]

julia> srv1 = Surv([1, 2, 3], [4, 5, 6])
[(1.0, 4.0), (2.0, 5.0), (3.0, 6.0)]

julia> srv2 = Surv([7, 8, 9], [10, 11, 12])
[(7.0, 10.0), (8.0, 11.0), (9.0, 12.0)]

julia> merge(srv1, srv2) # TwoSidedSurv
[(1.0, 4.0), (2.0, 5.0), (3.0, 6.0), (7.0, 10.0), (8.0, 11.0), (9.0, 12.0)]

Base.reverse — Method

reverse(srv::OneSidedSurv)

Reverse survival outcome (not in-place). Useful for computing non-parametric estimators of the censoring distribution instead of the survival distribution.

Examples

julia> srv = Surv([1, 2, 3, 4, 5, 6], [false, true, false, false, true, true], :r)
["1.0+", "2.0", "3.0+", "4.0+", "5.0", "6.0"]

julia> reverse(srv)
["1.0", "2.0+", "3.0", "4.0", "5.0+", "6.0+"]

SurvivalAnalysis.event_times — Method

outcome_times(srv::OneSidedSurv)

Get times at which an event (not censoring) takes place.

Examples

julia> event_times(Surv([1, 2, 3], [true, false, true], :r))
2-element Vector{Float64}:
 1.0
 3.0

SurvivalAnalysis.outcome_status — Method

outcome_status(srv::OneSidedSurv)

Get vector of survival indicators.

Examples

julia> outcome_status(Surv([1, 2, 3], [true, false, true], :r))
3-element Vector{Bool}:
 1
 0
 1

SurvivalAnalysis.outcome_times — Method

outcome_times(srv::OneSidedSurv)
outcome_times(srv::TwoSidedSurv)

Get times at which an outcome (event or censoring) takes place.

Examples

julia> outcome_times(Surv([1, 2, 3], [true, false, true], :r))
3-element Vector{Float64}:
 1.0
 2.0
 3.0

julia> outcome_times(Surv([1, 2, 3], [5, 6, 7]))
2-element Vector{Vector{Float64}}:
 [1.0, 2.0, 3.0]
 [5.0, 6.0, 7.0]

SurvivalAnalysis.surv_stats — Method

surv_stats(srv::OneSidedSurv; events_only = false)

Get set of useful summary statistics over time - if events_only = false (default) then returned for all outcomes, otherwise only return the times at which an event occurred (useful for computing non-parametric estimators).

Examples

julia> surv_stats(Surv([1, 2, 3], [false, true, false], :r))
(time = [1.0, 2.0, 3.0], nrisk = [3, 2, 1], ncens = [1, 0, 1], nevents = [0, 1, 0], noutcomes = [1, 1, 1])

julia> surv_stats(Surv([1, 2, 3], [false, true, false], :r); events_only = true)
(time = [2.0], nrisk = [2], ncens = [0], nevents = [1], noutcomes = [1])

SurvivalAnalysis.threshold_risk — Method

threshold_risk(srv::OneSidedSurv, p::Number)

Get the time at which a given proportion of observations are no longer at risk. Useful when computing metrics and calculations become increasingly unstable as number of observations at risk decreases over time.

Examples

# time at which 80% of observations have experienced the event or been censored
julia> threshold_risk(Surv([1, 2, 3, 4, 5, 6], [false, true, false, false, true, true], :r), 0.8)
6.0

SurvivalAnalysis.total_censored — Method

total_censored(srv::OneSidedSurv)
total_censored(srv::OneSidedSurv, t::Number)

Get total number censored (optionally at a given time).

Examples

julia> total_censored(Surv([1, 2, 3], [false, true, false], :r))
2

julia> total_censored(Surv([1, 2, 3], [true, false, true], :r), 3)
0

SurvivalAnalysis.total_events — Method

total_events(srv::OneSidedSurv)
total_events(srv::OneSidedSurv, t::Number)

Get total number of events (optionally at a given time).

Examples

julia> total_events(Surv([1, 2, 3], [true, false, true], :r))
2

julia> total_events(Surv([1, 2, 3], [true, false, true], :r), 3)
1

SurvivalAnalysis.total_outcomes — Method

total_outcomes(srv::OneSidedSurv)
total_outcomes(srv::OneSidedSurv, t::Number)

Get total number of outcomes (optionally at a given time).

Examples

julia> total_outcomes(Surv([1, 2, 3], [false, true, false], :r))
3

julia> total_outcomes(Surv([1, 2, 3], [true, false, true], :r), 3)
1

SurvivalAnalysis.total_risk — Method

total_risk(srv::OneSidedSurv)
total_risk(srv::OneSidedSurv, t::Number)

Get total number at risk (optionally at a given time).

Examples

julia> total_risk(Surv([1, 2, 3], [false, true, false], :r))
3

julia> total_risk(Surv([1, 2, 3], [true, false, true], :r), 2)
2

SurvivalAnalysis.unique_event_times — Method

unique_event_times(srv::OneSidedSurv)

Get unique times at which an event takes place.

Examples

julia> unique_event_times(Surv([1, 2, 1], [true, false, true], :r))
1-element Vector{Float64}:
 1.0

SurvivalAnalysis.unique_outcome_times — Method

unique_outcome_times(srv::OneSidedSurv)

Get unique times at which an outcome takes place

Examples

julia> unique_outcome_times(Surv([1, 2, 1], [true, false, true], :r))
2-element Vector{Float64}:
 1.0
 2.0

SurvivalAnalysis.Srv — Function

Srv(T::Symbol, Δ::Symbol, type::Int = 1)

Create a SurvTerm object for internal use for fitting models. Arguments, T and Δ should be a reference to the name of the time and status variables in a DataFrame respectively. type should be 1 (or omitted) for right-censoring or -1 for left-censoring.

Examples

julia> @formula(Srv(t, d) ~ 1) # right-censoring
FormulaTerm
Response:
  (t,d)->Srv(t, d)
Predictors:
  1

julia> @formula(Srv(t, d, -1) ~ X) # left-censoring
FormulaTerm
Response:
  (t,d)->Srv(t, d, -1)
Predictors:
  X(unknown)

Models

SurvivalEstimator

SurvivalAnalysis.KaplanMeier — Type

KaplanMeier <: SurvivalEstimator

See SurvivalEstimator.

SurvivalAnalysis.NelsonAalen — Type

NelsonAalen <: SurvivalEstimator

See SurvivalEstimator.

SurvivalAnalysis.SurvivalEstimator — Type

SurvivalEstimator <: StatisticalModel

Abstract type for all non-parametric estimators implemented in, or extending, this package. Type 'inherits' from StatsAPI.StatisticalModel to enable formula fitting and predicting interface. Note❗This may be abstracted further into ConditionalSurvivalEstimator <: SurvivalEstimator and UnconditionalSurvivalEstimator <: SurvivalEstimator.

Available methods:

fit and predict - Fit model and make predictions from fitted model with @formula or matrix interface, see Fitting and predicting

confint - Calculate confidence intervals around estimates

time - Extract fitted times

survival - Extract estimated survival probabilities

std - Extract computed standard deviation

distr - Extract fitted survival distribution

Objects inheriting from this should have the following fields:

time::Vector{Float64} - Fitted survival times

survival::Vector{Float64} - Estimated survival probabilities

std::Vector{Float64} - Computed standard deviation

distr::DiscreteNonParametric - Fitted survival distribution

stats::NamedTuple - Summary statistics such as numbers at risk, dead, censored

Base.Libc.time — Method

time(npe::SurvivalEstimator)
time(npe::StatsModels.TableStatisticalModel{<:SurvivalEstimator, <:AbstractMatrix})

Return fitted times from a SurvivalEstimator.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> time(km(@formula(Srv(Y, D) ~ 1), data))
4-element Vector{Float64}:
 1.0
 4.0
 5.0
 8.0

Statistics.std — Method

std(npe::SurvivalEstimator)
std(npe::StatsModels.TableStatisticalModel{<:SurvivalEstimator, <:AbstractMatrix})

Return computed standard deviation from a SurvivalEstimator.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> std(km(@formula(Srv(Y, D) ~ 1), data))
4-element Vector{Float64}:
 1.0004625991132958
 0.712458742539604
 0.6082063644330836
 0.5713145523891875

StatsAPI.confint — Method

confint(km::KaplanMeier; level::Float64 = 0.95)
confint(km::KaplanMeier, t::Number; level::Float64 = 0.95)

Calculate the confidence interval (CI) around a fitted Kaplan-Meier estimate to level% confidence. If t provided then returns CI at that time, otherwise returns CI at all fitted times. Standard deviation is calculated using the pointwise method of Kalbfleisch and Prentice (1980).

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> confint(km(@formula(Srv(Y, D) ~ 1), data))
4-element Vector{Tuple{Float64, Float64}}:
 (0.473009271362049, 0.9852813933673431)
 (0.38088152320549545, 0.9425909522237038)
 (0.21830025822743343, 0.8691223292427415)
 (0.0700802713627666, 0.7534316354804488)

StatsAPI.confint — Method

confint(na::NelsonAalen; level::Float64 = 0.95)
confint(na::NelsonAalen, t::Number; level::Float64 = 0.95)

Calculate the confidence interval (CI) around a fitted Nelson-Aalen estimate to level% confidence. If t provided then returns CI at that time, otherwise returns CI at all fitted times.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> confint(na(@formula(Srv(Y, D) ~ 1), data), 10)
(0.23186692958834798, 0.9464141496254767)

StatsAPI.fit — Method

fit(obj::Type{<:SurvivalEstimator}, X::AbstractMatrix{<:Real}, Y::RCSurv)
fit(obj::Type{<:SurvivalEstimator}, Y::RCSurv)

Fit a SurvivalEstimator survival model using matrix interface. It is recommended to use kaplan_meier or nelson_aalen directly instead.

StatsAPI.predict — Method

predict(fit::SurvivalEstimator, X::AbstractMatrix{<:Real})
predict(fit::StatsModels.TableStatisticalModel{SurvivalEstimator, Matrix{Float64}},
    data::DataFrames.DataFrame)

Make predictions from a fitted SurvivalEstimator model. See Fitting and predicting and examples below for predicting interfaces, we recommend using predict(fit, data::DataFrame).

Two prediction types can be made from a non-parametric estimator:

distr - See kaplan_meier or nelson_aalen for formulae

survival_matrix - Matrix of survival probabilities

Predicted distributions are returned as Distributions.DiscreteNonParametric.

Function returns a SurvivalPrediction struct.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> f = km(@formula(Srv(Y, D) ~ 1), data);

julia> predict(f, DataFrame(X = [2,9,1,1,exp(8),log(9),2^3,5^exp(2),1]));

SurvivalAnalysis.distr — Method

distr(npe::SurvivalEstimator)
distr(npe::StatsModels.TableStatisticalModel{<:SurvivalEstimator, <:AbstractMatrix})

Return estimated survival distribution from a SurvivalEstimator.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> distr(km(@formula(Srv(Y, D) ~ 1), data))
DiscreteNonParametric{Float64, Float64, Vector{Float64}, Vector{Float64}}(
support: [0.0, 1.0, 4.0, 5.0, 8.0]
p: [0.0, 0.11250000000000004, 0.15749999999999997, 0.20999999999999996]
)

SurvivalAnalysis.kaplan_meier — Method

kaplan_meier(Y::RCSurv)
kaplan_meier(f::@formula, data::DataFrames.DataFrame)
kaplan_meier(X::AbstractMatrix{<:Real}, Y::RCSurv)
fit(KaplanMeier, X::AbstractMatrix{<:Real}, Y::RCSurv)

Aliases: km, kaplan

Fit a non-parametric Kaplan-Meier estimator on a right-censored survival outcome. See Fitting and predicting and examples below for fitting interfaces, we recommend using kaplan_meier(::@formula...) (or aliases).

The Kaplan-Meier Estimator is defined by

\[Ŝ(τ) = ∏_{i:tᵢ≤τ} (1 - \frac{dᵢ}{nᵢ})\]

where $dᵢ$ and $nᵢ$ are the number of events and nunber at risk at time $tᵢ$ respectively.

Standard deviation is calculated using the pointwise method of Kalbfleisch and Prentice (1980).

Future additions:

Support for other censoring types (see #46)

Support for stratified estimator (see #47)

Function returns a KaplanMeier struct.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> f = kaplan_meier(@formula(Srv(Y, D) ~ 1), data)
StatsModels.TableStatisticalModel{KaplanMeier, Matrix{Float64}}

(Y,D;+) ~ 1

Coefficients:
  n  ncens  nevents
 10      6        4

SurvivalAnalysis.nelson_aalen — Method

nelson_aalen(Y::RCSurv)
nelson_aalen(f::@formula, data::DataFrames.DataFrame)
nelson_aalen(X::AbstractMatrix{<:Real}, Y::RCSurv)
fit(NelsonAalen, X::AbstractMatrix{<:Real}, Y::RCSurv)

Aliases: na, nelson

Fit a non-parametric Nelson-NelsonAalen estimator on a right-censored survival outcome. See Fitting and predicting and examples below for fitting interfaces, we recommend using nelson_aalen(::@formula...) (or aliases).

The Nelson-Aalen Estimator is defined by

\[Ĥ(τ) = ∑_{i:tᵢ≤τ} \frac{dᵢ}{nᵢ}\]

where $dᵢ$ and $nᵢ$ are the number of events and nunber at risk at time $tᵢ$ respectively and Ĥ is the estimated cumulative hazard function. The survival function is then Ŝ = exp(-Ĥ).

Future additions:

Support for other censoring types (see #46)

Support for stratified estimator (see #47)

Function returns a NelsonAalen struct.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> f = nelson_aalen(@formula(Srv(Y, D) ~ 1), data)
StatsModels.TableStatisticalModel{NelsonAalen, Matrix{Float64}}

(Y,D;+) ~ 1

Coefficients:
  n  ncens  nevents
 10      6        4

SurvivalAnalysis.survival — Method

survival(npe::SurvivalEstimator)
survival(npe::StatsModels.TableStatisticalModel{<:SurvivalEstimator, <:AbstractMatrix})

Return fitted survival probabilities from a SurvivalEstimator.

Examples

julia> data = DataFrame(Y = [1,1,4,6,8,4,9,4,5,10], D = [true, false, false, false, true, false, false, true, true, false]);

julia> survival(km(@formula(Srv(Y, D) ~ 1), data))
4-element Vector{Float64}:
 0.9
 0.7875
 0.63
 0.42000000000000004

SurvivalModel

SurvivalAnalysis.SurvivalModel — Type

SurvivalModel <: StatisticalModel

Abstract type for all models implemented in, or extending, this package. Type 'inherits' from StatsAPI.StatisticalModel to enable formula fitting and predicting interface.

ParametricSurvival

SurvivalAnalysis.ParametricAFT — Type

ParametricAFT{<:ContinuousUnivariateDistribution} <: ParametricSurvival

See ParametricSurvival.

SurvivalAnalysis.ParametricPH — Type

ParametricPH{<:ContinuousUnivariateDistribution} <: ParametricSurvival

See ParametricSurvival.

SurvivalAnalysis.ParametricSurvival — Type

ParametricSurvival <: SurvivalModel

Abstract type for all fully-parametric survival models implemented in, or extending, this package. Type 'inherits' SurvivalModel.

Available methods:

coef - Extract fitted coefficients

fit and predict - Fit model and make predictions from fitted model with @formula or matrix interface, see Fitting and predicting

baseline - Extract fitted baseline distribution, see ph and aft for more

scale - Extract scale parameter of fitted distribution

All distributions are fitted with the Kalbfleisch-Prentice parametrisation and then converted as required to make use of Distributions.jl

Objects inheriting from this should have the following fields:

coefficients::Vector{Float64} - Fitted coefficients

scale::Float64 - Fitted scale parameter for baseline distribution before transformation

hessian::Matrix - Hessian from Optim

var_cov::Matrix - Covariance matrix

tstats::Vector - t-statistics

baseline<:ContinuousUnivariateDistribution - Fitted baseline distribution

routine - Optimisation routine from Optim

Distributions.scale — Method

scale(mm::StatsModels.TableStatisticalModel{<:ParametricSurvival, <:AbstractMatrix})
scale(mm::ParametricSurvival)

Extract estimated scale parameter (before transformation) for baseline distribution from fitted ParametricSurvival object.

StatsAPI.fit — Method

fit(t::Type{<:ParametricSurvival}, X::AbstractMatrix{<:Real}, Y::RCSurv,
    d::Type{<:ContinuousUnivariateDistribution}; init::Number = 1)

Fit a ParametricSurvival survival model using matrix interface. It is recommended to use ph or aft directly instead.

StatsAPI.predict — Method

predict(fit::ParametricAFT, X::AbstractMatrix{<:Real})
predict(fit::StatsModels.TableStatisticalModel{ParametricAFT, Matrix{Float64}},
    data::DataFrames.DataFrame)

Make predictions from a fitted ParametricAFT model. See Fitting and predicting and examples below for predicting interfaces, we recommend using predict(fit, data::DataFrame).

Three prediction types can be made from a fitted AFT model:

lp - $Xβ̂$

crank - $-Xβ̂$

distr - $F(t) = F̂₀(t/exp(Xβ̂))$

where $β̂$ are estimated coefficients, $X$ are covariates from the new data, and $F̂₀$ is the estimated baseline distribution CDF function. Predicted distributions are returned as ContinuousAFTDistribution <: Distributions.ContinuousUnivariateDistribution.

Note❗The AFT model assumes that a higher linear predictor means a lower risk of event and therefore a higher survival time, i.e., $βXᵢ > βXⱼ → hᵢ(t) < hⱼ(t)$ - hence crank = -lp.

Future updates will add transformation methods for more prediction types (see #12).

Function returns a SurvivalPrediction struct.

Examples

julia> Y = [1,1,4,6,8,4,9,4,5,10];

julia> D = [true, false, false, false, true, false, false, true, true, false];

julia> X = [1,9,3,4,20,-4,pi,exp(5),log(8),0];

julia> data = DataFrame(Y = Y, D = D, X = X);

julia> f = aft(@formula(Srv(Y, D) ~ X), data, Exponential);

julia> predict(f, DataFrame(X = [2,9,1,1,exp(8),log(9),2^3,5^exp(2),1]));

StatsAPI.predict — Method

predict(fit::ParametricPH, X::AbstractMatrix{<:Real})
predict(fit::StatsModels.TableStatisticalModel{ParametricPH, Matrix{Float64}},
    data::DataFrames.DataFrame)

Make predictions from a fitted ParametricPH model. See Fitting and predicting and examples below for predicting interfaces, we recommend using predict(fit, data::DataFrame).

Three prediction types can be made from a fitted PH model:

lp - $Xβ̂$

crank - $Xβ̂$

distr - $F(t) = 1 - Ŝ₀(t)^{exp(Xβ̂)}$

where $β̂$ are estimated coefficients, $X$ are covariates from the new data, and $Ŝ₀$ is the estimated baseline distribution survival function. Predicted distributions are returned as ContinuousPHDistribution <: Distributions.ContinuousUnivariateDistribution.

Note❗The PH model assumes that a higher linear predictor means a higher risk of event and therefore a lower survival time, i.e., `βXᵢ > βXⱼ → hᵢ(t) > hⱼ(t) - hence crank = lp. This means when calculating concordance you must include rev = true.

Future updates will add transformation methods for more prediction types (see #12).

Function returns a SurvivalPrediction struct.

Examples

julia> Y = [1,1,4,6,8,4,9,4,5,10];

julia> D = [true, false, false, false, true, false, false, true, true, false];

julia> X = [1,9,3,4,20,-4,pi,exp(5),log(8),0];

julia> data = DataFrame(Y = Y, D = D, X = X);

julia> f = ph(@formula(Srv(Y, D) ~ X), data, Weibull);

julia> predict(f, DataFrame(X = [2,9,1,1,exp(8),log(9),2^3,5^exp(2),1]));

SurvivalAnalysis.aft — Method

aft(f::@formula, data::DataFrames.DataFrame, d::Type{T}; init::Number = 1)
aft(X::AbstractMatrix{<:Real}, Y::RCSurv, d::Type{T}; init::Number = 1)
fit(ParametricAFT, X::AbstractMatrix{<:Real}, Y::RCSurv, d::Type{T}; init::Number = 1)

Fit a fully-parametric accelerated failure time (AFT) model with baseline distribution d. See Fitting and predicting and examples below for fitting interfaces, we recommend using aft(::@formula...).

Fully-parametric AFT models are defined by

\[h(t) = e^{-Xβ} h₀(t e^{-Xβ})\]

where $β$ are coefficients to be estimated, $X$ are covariates, and h₀ is the hazard function of an assumed baseline distribution, d. Available choices for distributions are:

Distributions.Exponential

Distributions.Weibull

AFT models assumes that an increase in a covariate results in an acceleration of the event by a constant. This is best explained by example. The above formula can also be expressed as $S(t) = S₀(exp(-η)t)$ then let $ηᵢ = log(2)$ and $ηⱼ = log(1)$ so $ηᵢ > ηⱼ$. Then Sᵢ(t) = S₀(0.5t) and Sⱼ(t) = S₀(t) and so for all $t$, $Sᵢ(t) ≥ Sⱼ(t)$ as S is a decreasing function.

Future additions:

More baseline distributions (see #15)

Function returns a ParametricAFT struct.

Examples

julia> Y = [1,1,4,6,8,4,9,4,5,10];

julia> D = [true, false, false, false, true, false, false, true, true, false];

julia> X = [1,9,3,4,20,-4,2,4,1,0];

julia> data = DataFrame(Y = Y, D = D, X = X);

julia> f = aft(@formula(Srv(Y, D) ~ X), data, Weibull)
StatsModels.TableStatisticalModel{ParametricAFT{Weibull}, Matrix{Float64}}

(Y,D;+) ~ 1 + X

Distr:
Weibull{Float64}(α=1.491407, θ=11.636424)

Coefficients:
  (Scale)  (Intercept)           X
 0.670508      2.45414  -0.0209422

SurvivalAnalysis.baseline — Method

baseline(mm::StatsModels.TableStatisticalModel{<:ParametricSurvival, <:AbstractMatrix})
baseline(mm::ParametricSurvival)

Extract baseline distribution from fitted ParametricSurvival object.

SurvivalAnalysis.ph — Method

ph(f::@formula, data::DataFrames.DataFrame, d::Type{T}; init::Number = 1)
ph(X::AbstractMatrix{<:Real}, Y::RCSurv, d::Type{T}; init::Number = 1)
fit(ParametricPH, X::AbstractMatrix{<:Real}, Y::RCSurv, d::Type{T}; init::Number = 1)

Fit a fully-parametric proportional hazards (PH) model with baseline distribution d. See Fitting and predicting and examples below for fitting interfaces, we recommend using ph(::@formula...).

Fully-parametric PH models are defined by

\[h(t) = h₀(t)exp(Xβ)\]

where $β$ are coefficients to be estimated, $X$ are covariates, and h₀ is the hazard function of an assumed baseline distribution, d. Available choices for distributions are:

Distributions.Exponential

Distributions.Weibull

No other distributions have the PH assumption, which assumes that the risk of an event taking place is constant over time.

Future additions:

Methods for testing if the assumption is valid for your data (see #11).

Methods to calculate hazards ratios and add to show (see #40).

Function returns a ParametricPH struct.

Examples

julia> Y = [1,1,4,6,8,4,9,4,5,10];

julia> D = [true, false, false, false, true, false, false, true, true, false];

julia> X = [1,9,3,4,20,-4,pi,exp(5),log(8),0];

julia> data = DataFrame(Y = Y, D = D, X = X);

julia> f = ph(@formula(Srv(Y, D) ~ X), data, Exponential)
StatsModels.TableStatisticalModel{ParametricPH{Exponential}, Matrix{Float64}}

(Y,D;+) ~ 1 + X

Distr:
Exponential{Float64}(θ=17.121273)

Coefficients:
 (Scale)  (Intercept)          X
     1.0     -2.84032  0.0101247

SurvivalPrediction

SurvivalAnalysis.SurvivalPrediction — Method

SurvivalPrediction(;
    distr::Union{Nothing, Vector{<:Distribution}} = nothing,
    lp::Union{Nothing, Vector{T}}  = nothing,
    crank::Union{Nothing, Vector{T}} = nothing,
    time::Union{Nothing, Vector{T}} = nothing,
    fit_times::Union{Nothing, Vector{T}} = nothing,
    survival_matrix::Union{Matrix{T}, Nothing} = nothing
    ) where {T<:Number}

Survival models can make multiple types of predictions including:

distr - A survival time distribution (implemented with Distributions.jl)

lp - A linear predictor (usually Xβ, i.e. covariates * fitted coefficients)

crank - A generic continuous relative risk ranking

time - A survival time

survival_matrix - A matrix of predicted survival probabilities where rows are

observations and columns are fitted survival times corresponding to fit_times.

These predictions can only exist in a finite number of combinations so they are aggregated in types within this package (and automatically determined within this function):

DeterministicSurvivalPrediction(lp, crank, time)

DiscreteSurvivalPrediction(distr, lp, crank, time, survival_matrix)

ContinuousSurvivalPrediction(distr, lp, crank, time)

Absolutely no transformations take require assumptions take place within this function but it does transform survival_matrix to a Distributions.DiscreteNonParametric and vice versa.

Examples

julia> SurvivalPrediction(time = randn(5)); # DeterministicSurvivalPrediction

julia> SurvivalPrediction(fit_times = randn(5), survival_matrix = randn((2, 5))); # DiscreteSurvivalPrediction

julia> SurvivalPrediction(distr = fill(Exponential(), 2)); # ContinuousSurvivalPrediction

Measures

SurvivalAnalysis.SurvivalMeasure — Type

SurvivalMeasure

Abstract type for all measures implemented in, or extending, this package.

SurvivalAnalysis.MAE — Method

MAE(truth::OneSidedSurv, prediction::Vector{<:Number})

Calculates the Mean Absolute Error (MAE), defined by

\[MAEᵢ = δᵢ|tᵢ - t̂ᵢ| \ mean = \frac{1}{m} ∑ᵢ MAEᵢ \ se = \frac{σ}{√m}\]

where t is the true survival time, t̂ is the predicted survival time, δ is the censoring indicator, $m = ∑ᵢ δᵢ$, and σ is the sample standard deviation of MAE.

Returns a MAE struct inheriting from SurvivalTimeMeasure.

Note❗Censored observations in the test set are ignored.

Examples

julia> truth = Surv([1, 4, 2, 2, 8], [true, false, false, true, false], :r);

julia> pred = [pi, sqrt(3), 1, 8, 9];

julia> MAE(truth, pred)
MAE = 4.0708 (σ = 1.9292)

SurvivalAnalysis.MSE — Method

MSE(truth::OneSidedSurv, prediction::Vector{<:Number})

Calculates the Mean Squared Error (MSE), defined by

\[MSEᵢ = δᵢ(tᵢ - t̂ᵢ)^2 \ mean = \frac{1}{m} ∑ᵢ MSEᵢ \ se = \frac{σ}{√m}\]

where t is the true survival time, t̂ is the predicted survival time, δ is the censoring indicator, $m = ∑ᵢ δᵢ$, and σ is the sample standard deviation of MSE.

Returns a MSE struct inheriting from SurvivalTimeMeasure.

Note❗Censored observations in the test set are ignored.

Examples

julia> truth = Surv([1, 4, 2, 2, 8], [true, false, false, true, false], :r);

julia> pred = [pi, sqrt(3), 1, 8, 9];

julia> MSE(truth, pred)
MSE = 20.2932 (σ = 15.7068)

SurvivalAnalysis.RMSE — Method

RMSE(truth::OneSidedSurv, prediction::Vector{<:Number})

Calculates the Root Mean Squared Error (RMSE), defined by

\[MSEᵢ = δᵢ(tᵢ - t̂ᵢ)^2 \ RMSE = \sqrt{\frac{1}{m} ∑ᵢ MSEᵢ} \ se = σ/(2√m * rmse)\]

where t is the true survival time, t̂ is the predicted survival time, δ is the censoring indicator, $m = ∑ᵢ δᵢ$, and σ is the sample standard deviation of MSE.

Returns a RMSE struct inheriting from SurvivalTimeMeasure.

Note❗Censored observations in the test set are ignored. Note❗RMSE is an aggregated measure which means losses actually correspond to MSEᵢ.

Examples

julia> truth = Surv([1, 4, 2, 2, 8], [true, false, false, true, false], :r);

julia> pred = [pi, sqrt(3), 1, 8, 9];

julia> RMSE(truth, pred)
RMSE = 4.5048 (σ = 1.7433)

SurvivalAnalysis.SurvivalTimeMeasure — Type

SurvivalTimeMeasure

Abstract type for all survival time measures implemented in, or extending, this package. A survival time measure is any measure of the form $L(t, t̂)$ where t is the observed event time for an observation and t̂ is a prediction of the survival time.

Note❗Survival time measures are poorly researched in the literature and no good studies demonstrate their utility. Nor is there concensus on the 'right' way to implement these. Here the measures remove all censored observations and then compare predicted survival times for uncensored observations. Alternatives could include: using an IPCW weighting method, or retaining censoring observations and applying an uncertainty penalty.

Fields

Structs inheriting from this type include the following fields

losses - Pointwise losses for the measure, i.e. $L(tᵢ, t̂ᵢ)$

mean - Aggregation of losses, usually the sample mean

se - Standard error of the mean

Note❗ RMSE is an aggregated measure which means mean is not simply the sample mean over the losses and instead mean is just used as a shorthand.

SurvivalAnalysis.ConcordanceWeights — Type

ConcordanceWeights(S::Int8, G::Int8, tied_preds::Float64 tied_times::Float64)

Weights used in the concordance function. S and G reflect the power applied to the Kaplan-Meier estimates of the survival and censoring distributions of the fitted data used to apply weighting at a given time. For example ConcordanceWeights(1, -2, 0.5, 0.5) will tell the concordance function to multiply concordant pairs at time t by S(t)/G(t)². tied_preds and tied_times determine how to handle ties in predictions and observed times respectively.

See concordance for full examples.

Note❗It is strongly recommended that S ≥ 0 and G ≤ 0.

SurvivalAnalysis.concordance — Function

concordance(
    truth::OneSidedSurv, prediction::Vector{<:Number}, weights::Union{Symbol, ConcordanceWeights};
    tied_preds=0.5, tied_times=0, cutoff=nothing, train::OneSidedSurv=nothing, rev=false
)

Aliases: cindex

Generic function to call any concordance index method. Concordance is a measure of discrimination which evaluates if a prediction is concordant with the truth, i.e. let Tᵢ,Tⱼ be the true survival times, truth, for observations i and j and let ϕᵢ, ϕⱼ be predictions, prediction, then these are concordant if $ϕᵢ > ϕⱼ ⟺ Tᵢ > Tⱼ$.

In survival analysis the generic C-index is defined as follows

\[C = \frac{\sum_{i≠j} W(tᵢ)I(tᵢ < tⱼ, yᵢ < yⱼ, tᵢ < τ)δᵢ}{W(tᵢ)I(tᵢ < tⱼ, tᵢ < τ)δᵢ}\]

where tᵢ,tⱼ are true survival times, yᵢ,yⱼ are predictions, τ is a cutoff time used to ensure stability even when censoring is high, δᵢ is the censoring indicator and W is a weighting function determined by weights as follows:

:I or :Harrell - Harrell's C - $W(tᵢ) = 1$

:G2 or :Uno - Uno's C - $W(tᵢ) = 1/G(tᵢ)$

:SG or :Schemper - Schemper's C - $W(tᵢ) = S(tᵢ)/G(tᵢ)$

:S or :Peto - Peto-Wilcoxon's C - $W(tᵢ) = S(tᵢ)$

where S(tᵢ) and G(tᵢ) are respectively the Kaplan-Meier estimates of the survival and censoring distributions of the training data at time tᵢ. For any other combination of weights pass a ConcordanceWeights object to weights. Note❗If training data is not provided to train then truth is used to estimate S and G but it is strongly recommended to provide the training data if possibe.

We also include an implementation of Gönen-Heller's C with weights = :GH or :Gonen. Note❗This is actually a very different method from the others and calculates concordance for predictions from a Cox PH model only. We may move this to its own function in the future to avoid misuse.

There is open debate about how to handle ties when calculating the concordance. Ties can occur in predictions and in the observed survival times. The defaults here tied_preds=0.5 and tied_times=0 are set as these seem to be the most common but they can be changed. Note❗If you pass a ConcordanceWeights object then the tied weights specified in this will take priority.

Note❗For predictions from PH models or any model where the prediction represents a relative risk then a higher value of ϕ implies a higher risk of event which will result in a lower survival time. In this case a prediction is concordant with the survival time if $ϕᵢ < ϕⱼ ⟺ Tᵢ > Tⱼ$. To do this within the function just set rev=true.

Examples

julia> T = [1.0,0.1,pi,0.9,0.4,20,1,5,9,2.5];

julia> ϕ = [0.1,0.2,0.1,0.9,0.25,exp(2),8,log(9),2,8];

julia> Δ = [true,true,false,true,true,false,true,false,true,true];

julia> Y = Surv(T, Δ, :r);

julia> train = Surv([3,3,9,12,2,1,8,4,10,8], Δ, :r);

julia> concordance(Y, ϕ, cutoff = threshold_risk(Y, 0.8)) # Harrell's C cutoff when 80% data is censored or dead
SurvivalAnalysis.Concordance

Harrell's C = 0.6153846153846154

Cutoff: T ≤ 9.0
Counts:
 Pairs  Comparable  Concordant  Disconcordant
    90          39          23             14
Weights:
 IPCW  Tied preds  Tied times
    1         0.5         0.0
Ties:
 Times  Preds  Both
     1      2     0
Weighted calculation:
 Numerator  Denominator         C
      24.0         39.0  0.615385

julia> concordance(Y, ϕ, :Uno, train=train) # Uno's C
SurvivalAnalysis.Concordance

Uno's C = 0.6221374045801525

Cutoff: T ≤ 20.0
Counts:
 Pairs  Comparable  Concordant  Disconcordant
    90          39          23             14
Weights:
 IPCW  Tied preds  Tied times
 1/G²         0.5         0.0
Ties:
 Times  Preds  Both
     1      2     0
Weighted calculation:
 Numerator  Denominator         C
   28.1728       45.284  0.622137

julia> cindex(Y, ϕ, ConcordanceWeights(5, -3, 0.5, 0.5, "Silly Weights"); train=train) # Custom weights
SurvivalAnalysis.Concordance

Silly Weights C = 0.6070334788405888

Cutoff: T ≤ 20.0
Counts:
 Pairs  Comparable  Concordant  Disconcordant
    90          40          23             14
Weights:
  IPCW  Tied preds  Tied times
 S⁵/G³         0.5         0.5
Ties:
 Times  Preds  Both
     1      2     0
Weighted calculation:
 Numerator  Denominator         C
   25.6265       42.216  0.607033

Plotting

RecipesBase.apply_recipe — Function

plot(npe::SurvivalEstimator, plot_confint::Bool = true; level = 0.95)
plot(npe::StatsModels.TableStatisticalModel{SurvivalEstimator, Matrix{Float64}},
    plot_confint::Bool = true; level = 0.95)

Recipe for plotting fitted non-parametric estimators, npe. If plot_confint then confidence intervals also plotted at a level% confidence level.

Examples

julia> using Plots

julia> data = DataFrame(t = randn(10), d = [trues(5)..., falses(5)...]);

julia> plot(kaplan_meier(@formula(Srv(t, d) ~ 1), data));

julia> plot(nelson_aalen(@formula(Srv(t, d) ~ 1), data).model);

Tools

SurvivalAnalysis.cum_hazard — Method

cum_hazard(d::UnivariateDistribution, x::Real)

Aliases: Hₜ

Compute the cumulative hazard function of distribution d at point x.

The cumulative hazard function for random variable $t$ is defined as

\[H_t(x) = \int^x_0 h_t(u) du = -log(S_t(x))\]

where $hₜ$ is the hazard function of $t$ and $Sₜ$ is the survival function of $t$.

Examples

julia> using Distributions

julia> cum_hazard(Binomial(5, 0.5), 3)
1.6739764335716711

SurvivalAnalysis.hazard — Method

hazard(d::UnivariateDistribution, x::Real)

Aliases: hₜ

Compute the hazard function of distribution d at point x.

The hazard function for random variable $t$ is defined as

\[h_t(x) = \frac{f_t(x)}{S_t(x)}\]

where $fₜ$ is the pdf of $t$ and $Sₜ$ is the survival function of $t$.

Examples

julia> using Distributions

julia> hazard(Binomial(5, 0.5), 3)
1.6666666666666679

Index

SurvivalAnalysis.ConcordanceWeights

SurvivalAnalysis.KaplanMeier

SurvivalAnalysis.MAE

SurvivalAnalysis.MSE

SurvivalAnalysis.NelsonAalen

SurvivalAnalysis.ParametricAFT

SurvivalAnalysis.ParametricPH

SurvivalAnalysis.ParametricSurvival

SurvivalAnalysis.RMSE

SurvivalAnalysis.Surv

SurvivalAnalysis.Surv

SurvivalAnalysis.SurvivalEstimator

SurvivalAnalysis.SurvivalMeasure

SurvivalAnalysis.SurvivalModel

SurvivalAnalysis.SurvivalPrediction

SurvivalAnalysis.SurvivalTimeMeasure

Base.Libc.time

Base.length

Base.merge

Base.reverse

Distributions.scale

RecipesBase.apply_recipe

Statistics.std

StatsAPI.confint

StatsAPI.confint

StatsAPI.fit

StatsAPI.fit

StatsAPI.predict

StatsAPI.predict

StatsAPI.predict

SurvivalAnalysis.Srv

SurvivalAnalysis.aft

SurvivalAnalysis.baseline

SurvivalAnalysis.concordance

SurvivalAnalysis.cum_hazard

SurvivalAnalysis.distr

SurvivalAnalysis.event_times

SurvivalAnalysis.hazard

SurvivalAnalysis.kaplan_meier

SurvivalAnalysis.nelson_aalen

SurvivalAnalysis.outcome_status

SurvivalAnalysis.outcome_times

SurvivalAnalysis.ph

SurvivalAnalysis.surv_stats

SurvivalAnalysis.survival

SurvivalAnalysis.threshold_risk

SurvivalAnalysis.total_censored

SurvivalAnalysis.total_events

SurvivalAnalysis.total_outcomes

SurvivalAnalysis.total_risk

SurvivalAnalysis.unique_event_times

SurvivalAnalysis.unique_outcome_times